feat: move surprise me button to header
